Split HomeScreen in multiple files. Add new GetSavedFontFamilyItemsUseCase. Rename HomeUiState to HomeState and a new HomeUiState to handle UI updates.
