+++ app/src/main/java/dev/sergiobelda/foundry/di/UseCaseModule.kt
@@ -18,8 +18,9 @@ package dev.sergiobelda.foundry.di
 
 import dev.sergiobelda.foundry.domain.usecase.FetchFontsUseCase
 import dev.sergiobelda.foundry.domain.usecase.GetFontFamilyItemsUseCase
-import dev.sergiobelda.foundry.domain.usecase.SaveFontUseCase
+import dev.sergiobelda.foundry.domain.usecase.GetSavedFontFamilyItemsUseCase
 import dev.sergiobelda.foundry.domain.usecase.RemoveSavedFontUseCase
+import dev.sergiobelda.foundry.domain.usecase.SaveFontUseCase
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 
@@ -27,6 +28,7 @@ val useCaseModule =
     module {
         singleOf(::FetchFontsUseCase)
         singleOf(::GetFontFamilyItemsUseCase)
+        singleOf(::GetSavedFontFamilyItemsUseCase)
         singleOf(::SaveFontUseCase)
         singleOf(::RemoveSavedFontUseCase)
     }
