+++ app/src/main/java/dev/sergiobelda/foundry/domain/repository/IFontRepository.kt
@@ -28,5 +28,5 @@ interface IFontRepository {
 
     fun getFontFamilyItems(): Flow<List<FontFamilyItemModel>>
 
-    fun getSavedFontItems(): Flow<List<FontFamilyItemModel>>
+    fun getSavedFontFamilyItems(): Flow<List<FontFamilyItemModel>>
 }
