+++ app/src/main/java/dev/sergiobelda/foundry/ui/home/HomeUiState.kt
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2024 Sergio Belda
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dev.sergiobelda.foundry.ui.home
+
+import androidx.compose.material3.DrawerState
+import androidx.compose.material3.DrawerValue
+import androidx.compose.material3.rememberDrawerState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Stable
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
+
+@Composable
+fun rememberHomeUiState(
+    drawerState: DrawerState = rememberDrawerState(DrawerValue.Closed)
+): HomeUiState {
+    val coroutineScope = rememberCoroutineScope()
+    return remember(coroutineScope, drawerState) {
+        HomeUiState(
+            coroutineScope = coroutineScope,
+            drawerState = drawerState
+        )
+    }
+}
+
+@Stable
+class HomeUiState internal constructor(
+    private val coroutineScope: CoroutineScope,
+    val drawerState: DrawerState
+) {
+
+    fun openDrawer() {
+        coroutineScope.launch { drawerState.open() }
+    }
+
+    fun closeDrawer() {
+        coroutineScope.launch { drawerState.close() }
+    }
+}
